# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Desafio Compass",
    "description": "",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3333/api/v1"
    }
  ],
...
  "paths": {
    "/movies": {
      "get": {
        "tags": ["movie"],
        "summary": "Find all movies",
        "operationId": "findAllMovies",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MovieResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid Status Value"
          }
        }
      },
      "post": {
        "tags": ["movie"],
        "summary": "Creates a movie",
        "operationId": "createMovie",
        "requestBody": {
          "description": "JSON Example",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/movies/{id}": {
      "get": {
        "tags": ["movie"],
        "summary": "Find movie by ID",
        "description": "Returns a single movie",
        "operationId": "getMovieById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of movie to return",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["movie"],
        "summary": "Updates a movie",
        "operationId": "updateMovie",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of movie to update",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "JSON Example",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": ["movie"],
        "summary": "Deletes a movie",
        "operationId": "deleteMovie",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of movie to delete",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "no content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "MovieReq": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string",
            "example": "url_da_imagem"
          },
          "name": {
            "type": "string",
            "example": "nome_do_filme"
          },
          "description": {
            "type": "string",
            "example": "descricao_do_filme"
          },
          "actors": {
            "type": "string",
            "example": ["brad pit", "Ryan Goslin"]
          },
          "genre": {
            "type": "string",
            "example": "genero_do_filme"
          },
          "release_date": {
            "type": "string",
            "example": "03/06/2024"
          }
        },
        "required": [
          "image",
          "name",
          "description",
          "actors",
          "genre",
          "release_date"
        ]
      },
      "MovieResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "image": {
            "type": "string",
            "example": "url_da_imagem"
          },
          "name": {
            "type": "string",
            "example": "nome_do_filme"
          },
          "description": {
            "type": "string",
            "example": "descricao_do_filme"
          },
          "actors": {
            "type": "string",
            "example": ["Brad Pitt", "Ryan Gosling"]
          },
          "genre": {
            "type": "string",
            "example": "genero_do_filme"
          },
          "release_date": {
            "type": "string",
            "example": "03/06/2024"
          },
          "sessions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": []
          }
        },
        "required": [
          "id",
          "image",
          "name",
          "description",
          "actors",
          "genre",
          "release_date"
        ]
      },
...
    }
  }
}